Close AG.13 Group 4 self-IP proof
