+++ .triage/phase-AG/findings/AG13-FIND-B6-LOOPBACK-RELABELED-SELFIP.ttl
@@ -12,7 +12,7 @@
   triage:sweepScope "file_only" ;
   triage:category "test-coverage" ;
   triage:severity "blocking" ;
-  triage:status "open" ;
+  triage:status "closed" ;
   triage:dispatchReady true ;
   triage:triagedAt "2026-07-17T00:00:00Z"^^xsd:dateTime ;
   triage:hasOccurrence <urn:atm:triage:occurrence/AG13-FIND-B6/dispatcher-self-ip-send/480> ;
@@ -22,15 +22,15 @@
   triage:hasOccurrence <urn:atm:triage:occurrence/AG13-FIND-B6/loopback-template/1> ;
   triage:openOn <urn:atm:triage:worktree/pAG-s13-selfip-proof/6d2afb4c> ;
   triage:promoteTo <urn:atm:triage:worktree/pAG-s13-selfip-proof/6d2afb4c> ;
-  triage:notes "Sprint AG.13 core, still-unresolved defect: tests must bind/parse against the host's actual advertised/bound network-interface IP, not 127.0.0.1/localhost, to satisfy the sprint's named deliverable 'Self-IP proof'. Two test functions (dispatcher_self_ip_send_rejects_disabled_host_before_mailbox_mutation and dispatcher_secure_self_ip_requires_ack_round_trips_and_updates_reply_state) are currently scaffolded with loopback-only fixtures identical to pre-existing loopback tests. No fixture exercises real interface IP. Structural duplication confirmed by RBP review." .
+  triage:notes "Closed in AG.13 Group 4. The self-IP proof now uses a shared non-loopback self-IP fixture discovered at runtime (`192.0.0.2` in the validated run), the hardcoded `127.0.0.1` remote-host targets are gone from the self-IP tests, and the old loopback-relabeled unauthorized self-IP test was removed from runtime_root/loopback.rs. Validation: cargo fmt --all, just lint, cargo test -p atm-daemon -p agent-team-mail-core all passed." .
 
 <urn:atm:triage:occurrence/AG13-FIND-B6/dispatcher-self-ip-send/480>
   a triage:Occurrence ;
   triage:file "crates/atm-daemon/src/tests/runtime_root/loopback.rs" ;
   triage:line 480 ;
   triage:snippet "dispatcher_self_ip_send_rejects_disabled_host_before_mailbox_mutation() function definition [uses loopback-only fixtures]" ;
-  triage:status "open" ;
-  triage:closed false ;
+  triage:status "closed" ;
+  triage:closed true ;
   triage:branch "feature/pAG-s13-selfip-proof" ;
   triage:headSha "6d2afb4c" ;
   triage:occursIn <urn:atm:triage:worktree/pAG-s13-selfip-proof/6d2afb4c> .
@@ -40,8 +40,8 @@
   triage:file "crates/atm-daemon/src/tests/runtime_root/loopback.rs" ;
   triage:line 549 ;
   triage:snippet "RemoteTargetHost::parse(\"127.0.0.1\") [hardcoded loopback in self-IP test]" ;
-  triage:status "open" ;
-  triage:closed false ;
+  triage:status "closed" ;
+  triage:closed true ;
   triage:branch "feature/pAG-s13-selfip-proof" ;
   triage:headSha "6d2afb4c" ;
   triage:occursIn <urn:atm:triage:worktree/pAG-s13-selfip-proof/6d2afb4c> .
@@ -51,8 +51,8 @@
   triage:file "crates/atm-daemon/src/tests/runtime_root/loopback.rs" ;
   triage:line 590 ;
   triage:snippet "dispatcher_secure_self_ip_requires_ack_round_trips_and_updates_reply_state() function definition [uses loopback-only fixtures]" ;
-  triage:status "open" ;
-  triage:closed false ;
+  triage:status "closed" ;
+  triage:closed true ;
   triage:branch "feature/pAG-s13-selfip-proof" ;
   triage:headSha "6d2afb4c" ;
   triage:occursIn <urn:atm:triage:worktree/pAG-s13-selfip-proof/6d2afb4c> .
@@ -62,8 +62,8 @@
   triage:file "crates/atm-daemon/src/tests/runtime_root/loopback.rs" ;
   triage:line 646 ;
   triage:snippet "RemoteTargetHost::parse(\"127.0.0.1\") [hardcoded loopback in secure self-IP test]" ;
-  triage:status "open" ;
-  triage:closed false ;
+  triage:status "closed" ;
+  triage:closed true ;
   triage:branch "feature/pAG-s13-selfip-proof" ;
   triage:headSha "6d2afb4c" ;
   triage:occursIn <urn:atm:triage:worktree/pAG-s13-selfip-proof/6d2afb4c> .
@@ -73,8 +73,8 @@
   triage:file "crates/atm-daemon/src/tests/runtime_root/loopback.rs" ;
   triage:line 1 ;
   triage:snippet "dispatcher_loopback_send_round_trips_through_peer_listener_into_self_inbox() lines 1-93 [template for both self-IP tests: same tempdir/roster/allow_host scaffold]" ;
-  triage:status "open" ;
-  triage:closed false ;
+  triage:status "closed" ;
+  triage:closed true ;
   triage:branch "feature/pAG-s13-selfip-proof" ;
   triage:headSha "6d2afb4c" ;
   triage:occursIn <urn:atm:triage:worktree/pAG-s13-selfip-proof/6d2afb4c> .
