+++ crates/atm-daemon/src/tests/runtime_root/loopback.rs
@@ -485,118 +485,6 @@ fn dispatcher_secure_loopback_send_round_trips_through_peer_listener_into_self_i
 
 #[test]
 #[serial_test::serial(env)]
-fn dispatcher_self_ip_send_rejects_disabled_host_before_mailbox_mutation() {
-    install_retained_runtime_factory();
-    let tempdir = TempDir::new().expect("tempdir");
-    let atm_home = tempdir.path().join("atm-home");
-    let workspace_dir = tempdir.path().join("workspace");
-    std::fs::create_dir_all(&atm_home).expect("atm home dir");
-    std::fs::create_dir_all(&workspace_dir).expect("workspace dir");
-    let db_path = tempdir.path().join("mail.db");
-    write_team_config(&atm_home, &[]);
-
-    add_member_via_retained_admin(
-        &db_path,
-        &atm_home,
-        TEST_TEAM,
-        ROLE_TEAM_LEAD,
-        &workspace_dir,
-    );
-    add_member_via_retained_admin(&db_path, &atm_home, TEST_TEAM, "qa-a", &workspace_dir);
-
-    let assembly = open_sqlite_boundary(&db_path).expect("sqlite boundary");
-    let allowed_host = "127.0.0.1"
-        .parse::<atm_storage::AllowedHostName>()
-        .expect("allowed host");
-    assembly
-        .allowed_host_store_arc()
-        .allow_host(
-            atm_storage::AllowHostCommand::new(
-                "127.0.0.1",
-                format!("{ROLE_TEAM_LEAD}@{TEST_TEAM}"),
-                Some("self-ip rejection fixture".to_string()),
-            )
-            .expect("allow host command"),
-        )
-        .expect("allow host");
-    assembly
-        .allowed_host_store_arc()
-        .deny_host(&allowed_host)
-        .expect("deny host");
-
-    let status_cache = RuntimeStatusCache::new();
-    let peer_transport = crate::PeerTransportRuntime::new_server_for_test_with_allowed_host_store(
-        SocketAddr::from((std::net::Ipv4Addr::LOCALHOST, 0)),
-        crate::SubsystemObservability::disabled(crate::DaemonSubsystem::PeerTransport),
-        status_cache.clone(),
-        assembly.allowed_host_store_arc(),
-    );
-    let dispatcher = Arc::new(DaemonRequestDispatcher::new_for_test_with_peer_transport(
-        atm_home.clone(),
-        status_cache,
-        db_path.clone(),
-        peer_transport.clone(),
-    ));
-    peer_transport
-        .start(dispatcher.clone())
-        .expect("start peer listener");
-
-    let mut request = SendRequest::new(
-        atm_home.clone(),
-        workspace_dir.clone(),
-        ROLE_TEAM_LEAD.parse().expect("caller"),
-        "qa-a@test-team",
-        TEST_TEAM.parse().expect("team"),
-        SendMessageSource::Inline("unauthorized self ip".to_string()),
-        None,
-        false,
-        None,
-        false,
-    )
-    .expect("send request");
-    request.remote_host = atm_core::send::parse_send_target("qa-a@test-team.127.0.0.1", None)
-        .expect("parse target")
-        .remote_host;
-
-    let error = dispatcher
-        .dispatch(RequestEnvelope::Send(SendRequestEnvelope::Compose(request)))
-        .expect_err("self-ip send must fail closed");
-    assert_eq!(error.code, AtmErrorCode::MessageValidationFailed);
-
-    let read = dispatcher
-        .dispatch(RequestEnvelope::Receive(
-            ReadQuery::new(
-                atm_home,
-                workspace_dir,
-                "qa-a".parse().expect("caller"),
-                None,
-                TEST_TEAM.parse().expect("team"),
-                ReadSelection::All,
-                false,
-                false,
-                None,
-                None,
-                None,
-                None,
-                None,
-                None,
-            )
-            .expect("read query"),
-        ))
-        .expect("read self inbox");
-    let ResponseEnvelope::Receive(report) = read else {
-        panic!("expected receive response");
-    };
-    assert!(
-        report.message.is_none(),
-        "unauthorized self-ip send mutated the receiver mailbox"
-    );
-
-    peer_transport.shutdown().expect("shutdown peer listener");
-}
-
-#[test]
-#[serial_test::serial(env)]
 fn local_ipc_runtime_round_trips_send_after_add_member_roster_state() {
     install_retained_runtime_factory();
     let tempdir = TempDir::new().expect("tempdir");
