+++ crates/atm-daemon/src/tests/runtime_root/mod.rs
@@ -14,7 +14,7 @@ use atm_core::types::ReadSelection;
 use atm_runtime_test_support::{SQLITE_RUNTIME_PATH_ENV, open_sqlite_boundary};
 use atm_storage::{PeerSecurityMode, SetPeerSecurityModeCommand, UpsertTrustedPeerCommand};
 use std::io::Write;
-use std::net::SocketAddr;
+use std::net::{Ipv4Addr, SocketAddr, UdpSocket};
 use std::sync::Arc;
 use std::sync::mpsc;
 use std::time::Duration;
@@ -91,4 +91,20 @@ pub(super) fn configure_secure_loopback(db_path: &std::path::Path, host: &str) {
             .expect("trusted peer command"),
         )
         .expect("upsert trusted peer");
+}
+
+pub(super) fn discover_non_loopback_ipv4_for_test() -> Ipv4Addr {
+    let socket = UdpSocket::bind(SocketAddr::from((Ipv4Addr::UNSPECIFIED, 0)))
+        .expect("bind udp socket for self-ip discovery");
+    socket
+        .connect(SocketAddr::from(([198, 51, 100, 1], 9)))
+        .expect("connect udp socket for self-ip discovery");
+    match socket
+        .local_addr()
+        .expect("local addr for self-ip discovery")
+        .ip()
+    {
+        std::net::IpAddr::V4(ip) if !ip.is_loopback() && !ip.is_unspecified() => ip,
+        other => panic!("expected non-loopback local IPv4 for self-ip fixture, got {other}"),
+    }
 }
